+++ extensions/iceworks-ui-builder/web/src/pages/ComponentCreator/index.tsx
@@ -29,9 +29,12 @@ const Home = () => {
         content: intl.formatMessage({ id: 'web.iceworksUIBuilder.getMaterialError' }),
       });
     }
-
-    console.log('getSources', sources);
     return sources;
+  }
+
+  async function refreshSources() {
+    await callService('material', 'cleanCache');
+    return await getSources();
   }
 
   async function getData(source: string) {
@@ -139,6 +142,7 @@ const Home = () => {
             <Material
               disableLazyLoad
               getSources={getSources}
+              refreshSources={refreshSources}
               onSettingsClick={onSettingsClick}
               getData={getData}
               onBlockClick={onSelect}