+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.3.12
+
+- fix: show welcome page many times
+- fix: refresh material did not clean cache
+
 ## 0.3.11
 
 - feat: Add welcome page