+++ extensions/iceworks-style-helper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.3.9
+
+- feat: Add welcome page
+
 ## 0.3.7
 
 - Fix error: property `title` is mandatory and must be of type `string` or `object`.