+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.1.16
+
+- fix: show welcome page many times
+- fix: refresh material did not clean cache
+
 ## 0.1.15
 
 - feat: support generating menu to `menuConfig.js`